requirements.txt
fondant==0.10.1
pandas==2.2.0
pyarrow==15.0.0
dask==2024.2.0
tensorflow==2.15.0
biopython==1.83
ipykernel==6.29.3